更新一下
给N.镭加了一个形分
然后改了一下部分精灵的radius
好像没了

+++ SOURCE/LOGIC/SPRITE.py
@@ -17,7 +17,7 @@ class Base(pygame.sprite.Sprite):
         pos: tuple[int, int] = (0, 0),
         mask: bool = False,
         rotate: bool = False,
-        radius: int = None
+        radius: float = None
     ) -> None:
         super().__init__(*group)
